Major Code Refactor
